fix: convert unix time to long

+++ src/main/org/ctrip/ops/sysdev/utils/dateparser/UnixParser.java
@@ -6,7 +6,7 @@ public class UnixParser implements DateParser {
 
 	@Override
 	public DateTime parse(String input) {
-		return new DateTime(Double.parseDouble(input) * 1000);
+		return new DateTime((long) (Double.parseDouble(input) * 1000));
 	}
 
 	public static void main(String[] args) {
